lab3/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
Werkzeug==3.1.3
